Added Authentication
* Added SecurityConfiguration
* Added https
* Added UserRepository
* Added UserRepositoryAuthenticationProvider
* Added user examples

+++ appSpring/src/main/java/appSpring/repository/UserRepository.java
@@ -3,7 +3,7 @@ package appSpring.repository;
 import org.springframework.data.repository.CrudRepository;
 import appSpring.entity.User;
 
-public interface UserRepository extends CrudRepository<User, Long>{
+public interface UserRepository extends CrudRepository<User, Integer>{
 	
 	User findByName(String name);
 }
